pyproject.toml
[tool.poetry]
name = "first-3996-2023"
version = "0.1.0"
description = ""
authors = ["Martin Rioux"]
readme = "README.md"
packages = [{include = "first_3996_2023"}]

[tool.poetry.dependencies]
python = ">=3.7,<3.12"
robotpy-navx = "^2023.0.3"
robotpy-rev = "^2023.1.3.2"
robotpy-ctre = "^2023.0.0"
robotpy-wpilib-utilities = "2023.0.0"
robotpy = "^2023.3.2.1"
pyntcore = "^2023.3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
